YT-25879: Extract context_switch.h commit_hash:a4b3491c39f3c1675be244cf828aa3d4a9fe17e5
