+++ yt/yt/core/concurrency/scheduler_api.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "public.h"
+#include "context_switch.h"
 
 #include <yt/yt/core/actions/future.h>
 
@@ -37,55 +38,6 @@ TFiberId GetCurrentFiberId();
 void SetCurrentFiberId(TFiberId id);
 
 ////////////////////////////////////////////////////////////////////////////////
-
-//! Returns |true| if fiber context switch is currently forbidden.
-bool IsContextSwitchForbidden();
-
-class TForbidContextSwitchGuard
-{
-public:
-    TForbidContextSwitchGuard();
-    TForbidContextSwitchGuard(const TForbidContextSwitchGuard&) = delete;
-
-    ~TForbidContextSwitchGuard();
-
-private:
-    const bool OldValue_;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
-// NB: Use function pointer to minimize the overhead.
-using TGlobalContextSwitchHandler = void(*)();
-
-void InstallGlobalContextSwitchHandlers(
-    TGlobalContextSwitchHandler outHandler,
-    TGlobalContextSwitchHandler inHandler);
-
-////////////////////////////////////////////////////////////////////////////////
-
-using TContextSwitchHandler = std::function<void()>;
-
-class TContextSwitchGuard
-{
-public:
-    TContextSwitchGuard(
-        TContextSwitchHandler outHandler,
-        TContextSwitchHandler inHandler);
-
-    TContextSwitchGuard(const TContextSwitchGuard& other) = delete;
-    ~TContextSwitchGuard();
-};
-
-class TOneShotContextSwitchGuard
-    : public TContextSwitchGuard
-{
-public:
-    explicit TOneShotContextSwitchGuard(TContextSwitchHandler outHandler);
-
-private:
-    bool Active_;
-};
 
 //! Blocks the current fiber until #future is set.
 //! The fiber is resceduled to #invoker.
